requirements.txt
pymupdf==1.24.5
openai>=1.30.0
anthropic>=0.28.0
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0
Pillow>=10.0.0
chromadb>=0.5.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
